[33592] Allow type configuration to be empty
We currently signal resetting of attribute groups with an empty array.
If you wanted to remove all optional attributes and just be left with
subject and description, you couldn't do that at the moment.

This changes the reset signal to simple a single empty group, which
ensures the two things are handled differently.

https://community.openproject.com/wp/33592